#339 fix: Architecture clean up. Same format as main is now
