+++ apps/web/src/features/vms/utils/merge-vms.ts
@@ -1,0 +1,18 @@
+import { mockVms } from '@/__mocks__/data/vms'
+
+export const mergedVMs = (mockVms.resources || []).map((vm: any) => ({
+  ...vm,
+  metadata: {
+    ...vm.metadata,
+    creationtimestamp: {
+      ...vm.metadata.creationtimestamp,
+      time: vm.metadata.creationtimestamp?.time
+        ? {
+            date: {
+              numberLong: vm.metadata.creationtimestamp.time.$date?.$numberLong,
+            },
+          }
+        : undefined,
+    },
+  },
+}))
